requirements.txt
bcrypt==3.2.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.6
click==8.0.2
colorama==0.4.4
cryptography==35.0.0
cssselect2==0.4.1
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
flake8==4.0.1
Flask==2.0.2
Flask-Compress==1.10.1
Flask-WTF==0.15.1
future==0.18.2
graphviz==0.17
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
mccabe==0.6.1
netmiko==3.4.0
networkx==2.6.3
ntc-templates==2.3.2
paramiko==2.8.0
Pillow==8.4.0
plotly==5.3.1
pycodestyle==2.8.0
pycparser==2.20
pydot==1.4.2
pyflakes==2.4.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
reportlab==3.6.2
requests==2.26.0
schedule==1.1.0
scp==0.14.1
six==1.16.0
svglib==1.1.0
tabulate==0.8.9
tenacity==8.0.1
textfsm==1.1.2
tinycss2==1.1.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
webencodings==0.5.1
Werkzeug==2.0.2
WTForms==2.3.3
